hunyuan-ocr-service/requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
transformers
pillow
torch
